refactor: migrate backend to TypeScript
