+++ .eslintignore
@@ -1,0 +1,3 @@
+dist-backend/
+dist/
+node_modules/
