--- services/eventSubMessageTypes.js
@@ -1,7 +1,0 @@
-module.exports = {
-    NOTIFICATION: 'notification',
-    SESSION_WELCOME: 'session_welcome',
-    SESSION_KEEPALIVE: 'session_keepalive',
-    SESSION_RECONNECT: 'session_reconnect',
-    REVOCATION: 'revocation'
-};
