--- services/middleware/ActionTypes.js
@@ -1,10 +1,0 @@
-const ActionTypes = {
-    SEND_MESSAGE: 'SEND_MESSAGE',
-    MUTE_USER: 'MUTE_USER',
-};
-
-module.exports = {
-    ActionTypes,
-    SEND_MESSAGE: ActionTypes.SEND_MESSAGE,
-    MUTE_USER: ActionTypes.MUTE_USER,
-};
