--- services/middleware/Middleware.js
@@ -1,7 +1,0 @@
-class Middleware {
-    processMessage(message) {
-        throw new Error('processMessage not implemented');
-    }
-}
-
-module.exports = Middleware;
