+++ tsconfig.json
@@ -1,0 +1,21 @@
+{
+  "compilerOptions": {
+    "target": "ES2020",
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "esModuleInterop": true,
+    "resolveJsonModule": true,
+    "downlevelIteration": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "outDir": "dist-backend",
+    "noImplicitAny": false
+  },
+  "include": [
+    "main.ts",
+    "services/**/*.ts",
+    "windowsManager.ts",
+    "ipcHandlers.ts",
+    "webServer.ts"
+  ]
+}
